requirements.txt
# Core dependencies
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.115.0

# Data processing
python-dotenv==1.0.0
python-dateutil==2.8.2
beautifulsoup4==4.12.3
lxml==5.1.0

# Dashboard
streamlit==1.31.0
plotly==5.18.0
pandas==2.2.0

# Utilities
fuzzywuzzy==0.18.0
python-Levenshtein==0.25.0

# Testing
pytest==8.0.0
pytest-cov==4.1.0
